requirements.txt
protobuf==6.33.5
streamlit
pandas
openai
python-dotenv
google-adk>=1.24.0